requirements.txt
# ==============================
# Devia Backend AI Agent System
# ==============================

# --- Core AI / LLM ---
openai==1.35.0
semantic-kernel==1.0.3

# --- Web Framework & Server ---
fastapi==0.104.1
uvicorn[standard]==0.24.0

# --- Data Models & Validation ---
pydantic==2.5.0
pydantic-settings>=2.2.1,<3.0.0
typing-extensions==4.8.0

# --- Database ---
motor==3.3.2
pymongo==4.6.0
bson==0.5.10


# --- HTTP & Networking ---
httpx==0.25.2
requests==2.31.0

# --- Utilities ---
python-dotenv==1.0.0
python-dateutil==2.8.2
orjson==3.9.10
loguru==0.7.2

# --- UI & Visualization (for testing/demo) ---
streamlit==1.28.0
plotly==5.17.0
pandas==2.1.3

# --- Text Processing (optional) ---
nltk==3.8.1

# --- Dev / Testing Tools ---
black==23.11.0
flake8==6.1.0
isort==5.12.0
pytest==7.4.3
pytest-asyncio==0.21.1
